Sketchup Model Uploads
Added the "Hanging Garden Components" folder. This includes the basic components of the Hanging Garden - several iterations of them. Next iteration of this repository, I'll go through and do a bit of cleanup.

Also added the "Hanging Garden with Mods" file. This is a version of the design that includes some as-yet unbuilt components.